% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{baddeley.avg}
\alias{baddeley.avg}
\title{Compute aggregation of Ripley's K based on Baddeley et al. (1993)}
\usage{
baddeley.avg(K.vec, n.vec)
}
\arguments{
\item{K.vec}{Vector of Ripley's K estimates across a series of images}

\item{n.vec}{Vector of the number of points in each image}
}
\value{
Returns a weighted average of the Ripley's K values based on the number of points in each image
}
\description{
Compute aggregation of Ripley's K based on Baddeley et al. (1993)
}
\examples{
# Generate simulated Ripley's K values, areas, and number of points
K.vec <- rnorm(3, mean = 1000, sd = 100)
n.vec <- rnorm(3, mean = 300, sd = 50)

# Compute a weighted average
baddeley.avg(K.vec, n.vec)
}
\references{
Baddeley et al. (1993). Analysis of a Three-Dimensional Point Pattern with Replication.
Journal of the Royal Statistical Society. Series C (Applied Statistics), 42(4), 641-668.
}
