% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{diggle.avg}
\alias{diggle.avg}
\title{Compute aggregation of Ripley's K based on Diggle et al. (1991)}
\usage{
diggle.avg(K.vec, n.vec)
}
\arguments{
\item{K.vec}{Vector of Ripley's K estimates across a series of images}

\item{n.vec}{Vector of the number of points in each image}
}
\value{
Returns a weighted average of the Ripley's K values based on the number of points in each image
}
\description{
Compute aggregation of Ripley's K based on Diggle et al. (1991)
}
\examples{
# Generate simulated Ripley's K values, areas, and number of points
K.vec <- rnorm(3, mean = 1000, sd = 100)
n.vec <- rnorm(3, mean = 300, sd = 50)

# Compute a weighted average
diggle.avg(K.vec, n.vec)
}
\references{
Diggle et al. (1991). Analysis of Variance for Replicated Spatial Point Patterns in Clinical Neuroanatomy.
Journal of the American Statistical Association, 86(415), 618-625.
}
