% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagg-package.R
\docType{package}
\name{spagg-package}
\alias{spagg}
\alias{spagg-package}
\title{spagg: Aggregation approaches for combining spatial summary statistics}
\description{
Implements six aggregation approaches to combining spatial summary statistics estimated on multiple multiplexed spatial proteomics images or samples collected from the same location or person.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sarahsamorodnitsky.github.io/spagg/}
}

}
\keyword{internal}
